refactor: Move $renamedSettings & magic get/set to BaseConfig, add static $configCategory property
