+++ src/config/BaseConfig.php
@@ -8,7 +8,9 @@
 namespace craft\config;
 
 use craft\base\FluentModelTrait;
+use craft\services\Config;
 use yii\base\BaseObject;
+use yii\base\UnknownPropertyException;
 
 /**
  * Base config class
@@ -19,6 +21,17 @@ use yii\base\BaseObject;
 class BaseConfig extends BaseObject
 {
     use FluentModelTrait;
+
+    /**
+     * @var string The category for this config
+     */
+    public static $configCategory = '';
+
+    /**
+     * @var array Settings that have been renamed
+     */
+    protected static array $renamedSettings = [
+    ];
 
     /**
      * Factory method for creating new config objects
@@ -37,5 +50,50 @@ class BaseConfig extends BaseObject
     final public function __construct($config = [])
     {
         parent::__construct($config);
+    }
+    
+    /**
+     * @inheritdoc
+     */
+    public function __get($name)
+    {
+        if (isset(static::$renamedSettings[$name])) {
+            return $this->{static::$renamedSettings[$name]};
+        }
+
+        return parent::__get($name);
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function __set($name, $value)
+    {
+        if (isset(static::$renamedSettings[$name])) {
+            $newName = static::$renamedSettings[$name];
+            $configFilePath = Craft::$app->getConfig()->getConfigFilePath(static::$configCategory);
+            Craft::$app->getDeprecator()->log($name, "The `$name` config setting has been renamed to `$newName`.", $configFilePath);
+            $this->$newName = $value;
+            return;
+        }
+
+        try {
+            parent::__set($name, $value);
+        } catch (UnknownPropertyException) {
+            $category = static::$configCategory;
+            throw new UnknownPropertyException("Invalid $category config setting: $name. You can set custom config settings from config/custom.php.");
+        }
+    }
+
+    /**
+     * @inheritdoc
+     */
+    public function __isset($name)
+    {
+        if (isset(static::$renamedSettings[$name])) {
+            return isset($this->{static::$renamedSettings[$name]});
+        }
+
+        return parent::__isset($name);
     }
 }
